+++ cplib/is_prime.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <vector>
 #include "cplib/miller_rabin.hpp"
 
@@ -7,7 +8,7 @@ namespace cplib {
 
 // n が素数であるかを判定する(Miller-Rabin法)
 // n < 2^64, q <= 10^5 程度が間に合う
-bool is_prime(const unsigned long long n) {
+bool is_prime(const uint64_t n) {
     if (n < 4759123141ULL) {
         return miller_rabin(n, {2, 7, 61});
     } else {
